added RL struct for batch update

+++ src/RLFuzzing.py
@@ -46,26 +46,26 @@ class RLFuzzing:
             message, mtype = self.mq_reciever.receive()
 
             if mtype == INITALIZATION_FLAG:
-                self.map_size = int(np.frombuffer(message, dtype=np.uint64)[0])
+                self.map_size = int(np.frombuffer(message, dtype=np.uintc)[0])
                 print(f'self.map_size: {self.map_size}')
 
             elif mtype == UPDATE_SCORE:
                 print('1')
-                message_numpy_array = np.frombuffer(message, dtype=np.uint64)
+                message_numpy_array = np.frombuffer(message, dtype=np.uintc)
                 positive_reward = message_numpy_array
                 while len(positive_reward) < self.map_size:
                     print(f'len(positive_reward) : {len(positive_reward)}')
                     message, mtype = self.mq_reciever.receive()
-                    message_numpy_array = np.frombuffer(message, dtype=np.uint64)
+                    message_numpy_array = np.frombuffer(message, dtype=np.uintc)
                     positive_reward = np.concatenate([positive_reward, message_numpy_array])
 
                 print('2')
                 message, mtype = self.mq_reciever.receive()
-                message_numpy_array = np.frombuffer(message, dtype=np.uint64)
+                message_numpy_array = np.frombuffer(message, dtype=np.uintc)
                 negative_reward = message_numpy_array
                 while len(negative_reward) < self.map_size:
                     message, mtype = self.mq_reciever.receive()
-                    message_numpy_array = np.frombuffer(message, dtype=np.uint64)
+                    message_numpy_array = np.frombuffer(message, dtype=np.uintc)
                     negative_reward = np.concatenate([negative_reward, message_numpy_array])
 
                 self.positive_reward = positive_reward[:self.map_size]
@@ -106,7 +106,7 @@ class RLFuzzing:
             msg_npy = np.zeros(BUFF_SIZE_SENDER)
             msg_npy[0] = best_seed_id
             msg_npy[1] = self.positive_reward[best_seed_id] + self.negative_reward[best_seed_id]
-            msg_npy = np.array(msg_npy, dtype=np.uint64).reshape((2,BUFF_SIZE_SENDER//2))
+            msg_npy = np.array(msg_npy, dtype=np.uintc).reshape((2,BUFF_SIZE_SENDER//2))
             try:
                 self.mq_sender.send(msg_npy.tobytes(order='C'), True, type=mtype)
             except sysv_ipc.ExistentialError:
